requirements.txt
email-validator>=2.2.0
flask-dance>=7.1.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
pyjwt>=2.10.1
flask-login>=0.6.3
oauthlib>=3.3.1
reportlab>=4.4.2
requests>=2.32.4
whois>=1.20240129.2
sqlalchemy>=2.0.41
werkzeug>=3.1.3
python-whois>=0.9.5
pytz>=2025.2
dnspython>=2.7.0
scikit-learn>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
pymysql
flask-bcrypt
